Refactor account controller to use user model.
