+++ controller/account/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+    changePassword: require('./change-password'),
+    remove: require('./remove')
+};
